Upgrading To 12.0 From 11.x

+++ interex-web/app/Http/Controllers/Agent/Auth/ForgotPasswordController.php
@@ -150,7 +150,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return redirect()->route('agent.password.forgot')->with(['error' => [__('Password Reset Token Expired')]]);
         $resend_time = 0;
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
-            $resend_time = Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
+            $resend_time = (int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
         }
         $user_email = $password_reset->agent->email ?? "";
         return view('agent.auth.forgot-password.verify',compact('page_title','token','user_email','resend_time'));
@@ -208,7 +208,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return back()->with(['error' => [__('Password Reset Token Expired')]]);
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
             throw ValidationException::withMessages([
-                'code'      => __('You can resend the verification code after').' '.Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
+                'code'      => __('You can resend the verification code after').' '.(int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
             ]);
         }
 
@@ -278,7 +278,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return redirect()->route('agent.password.forgot')->with(['error' => [__('Password Reset Token Expired')]]);
         $resend_time = 0;
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
-            $resend_time = Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
+            $resend_time = (int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
         }
         $user_mobile = $password_reset->agent->full_mobile ?? "";
         return view('agent.auth.sms-password.verify',compact('page_title','token','user_mobile','resend_time'));
@@ -290,7 +290,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return back()->with(['error' => [__('Password Reset Token Expired')]]);
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
             throw ValidationException::withMessages([
-                'code'      => __('You can resend the verification code after').' '.Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
+                'code'      => __('You can resend the verification code after').' '.(int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
             ]);
         }
         DB::beginTransaction();
