+++ interex-web/app/Http/Controllers/Api/Merchant/AuthorizationController.php
@@ -39,7 +39,7 @@ class AuthorizationController extends Controller
         $resend = MerchantAuthorization::where("merchant_id",$user->id)->first();
         if( $resend){
             if(Carbon::now() <= $resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)) {
-                $error = ['error'=>[ __("You can resend the verification code after").' '.Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
+                $error = ['error'=>[ __("You can resend the verification code after").' '.(int) Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
                 return Helpers::error($error);
             }
         }
@@ -115,7 +115,7 @@ class AuthorizationController extends Controller
         $resend = MerchantAuthorization::where("merchant_id",$user->id)->first();
         if( $resend){
             if(Carbon::now() <= $resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)) {
-                $error = ['error'=>[ __("You can resend the verification code after").' '.Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
+                $error = ['error'=>[ __("You can resend the verification code after").' '.(int) Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
                 return Helpers::error($error);
             }
         }
@@ -509,7 +509,7 @@ class AuthorizationController extends Controller
         $resend = MerchantAuthorization::where("email",$request->email)->first();
         if($resend){
             if(Carbon::now() <= $resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)) {
-                $message = ['error'=>[__("You can resend the verification code after").' '.Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
+                $message = ['error'=>[__("You can resend the verification code after").' '.(int) Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
                 return Helpers::error($message);
             }
         }
@@ -611,7 +611,7 @@ class AuthorizationController extends Controller
         $resend = MerchantAuthorization::where("phone", $full_mobile)->first();
         if($resend){
             if(Carbon::now() <= $resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)) {
-                $message = ['error'=>[__("You can resend the verification code after").' '.Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
+                $message = ['error'=>[__("You can resend the verification code after").' '.(int) Carbon::now()->diffInSeconds($resend->created_at->addMinutes(GlobalConst::USER_VERIFY_RESEND_TIME_MINUTE)). ' '.__('seconds')]];
                 return Helpers::error($message);
             }
         }
