+++ interex-web/app/Http/Controllers/Merchant/Auth/ForgotPasswordController.php
@@ -149,7 +149,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return redirect()->route('merchant.password.forgot')->with(['error' => [__('Password Reset Token Expired')]]);
         $resend_time = 0;
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
-            $resend_time = Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
+            $resend_time = (int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
         }
         $user_email = $password_reset->merchant->email ?? "";
         return view('merchant.auth.forgot-password.verify',compact('page_title','token','user_email','resend_time'));
@@ -207,7 +207,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return back()->with(['error' => [__('Password Reset Token Expired')]]);
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
             throw ValidationException::withMessages([
-                'code'      => __('You can resend the verification code after').' '.Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
+                'code'      => __('You can resend the verification code after').' '.(int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
             ]);
         }
 
@@ -276,7 +276,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return redirect()->route('merchant.password.forgot')->with(['error' => [__('Password Reset Token Expired')]]);
         $resend_time = 0;
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
-            $resend_time = Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
+            $resend_time = (int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE));
         }
         $user_mobile = $password_reset->merchant->full_mobile ?? "";
         return view('merchant.auth.sms-password.verify',compact('page_title','token','user_mobile','resend_time'));
@@ -288,7 +288,7 @@ class ForgotPasswordController extends Controller
         if(!$password_reset) return back()->with(['error' => [__('Password Reset Token Expired')]]);
         if(Carbon::now() <= $password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)) {
             throw ValidationException::withMessages([
-                'code'      => __('You can resend the verification code after').' '.Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
+                'code'      => __('You can resend the verification code after').' '.(int) Carbon::now()->diffInSeconds($password_reset->created_at->addMinutes(GlobalConst::USER_PASS_RESEND_TIME_MINUTE)). ' '. __('seconds'),
             ]);
         }
         DB::beginTransaction();
